+++ src/main/java/com/addiis/core/gestionlogistica/persistence/repositories/ProductRepository.java
@@ -4,8 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.addiis.core.gestionlogistica.persistence.entities.product.Product;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    Optional<Product> findByEan(Integer ean);
+    Optional<Product> findByEan(BigInteger ean);
 }
